fixed more size_t to int comparisons

+++ Libft/ft_memmove.c
@@ -6,7 +6,7 @@
 /*   By: adeters <adeters@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/07 13:12:53 by adeters           #+#    #+#             */
-/*   Updated: 2024/09/07 14:34:43 by adeters          ###   ########.fr       */
+/*   Updated: 2024/09/08 16:14:41 by adeters          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@ void	*ft_memmove(void *dest, const void *src, size_t n)
 {
 	unsigned char	*dest_ptr;
 	unsigned char	*src_ptr;
-	int				i;
+	size_t			i;
 
 	i = 0;
 	dest_ptr = (unsigned char *)dest;
