+++ Libft/ft_strnstr.c
@@ -6,7 +6,7 @@
 /*   By: adeters <adeters@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/05 12:07:46 by adeters           #+#    #+#             */
-/*   Updated: 2024/09/05 12:08:15 by adeters          ###   ########.fr       */
+/*   Updated: 2024/09/08 16:19:37 by adeters          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strnstr(const char *big, const char *little, size_t len)
 {
-	int	i;
-	int	j;
+	size_t	i;
+	int		j;
 
 	i = 0;
 	if (little[i] == '\0')
